added some overwrite-preventing security mechanisms for the receiver

+++ src/receiver/handler.rs
@@ -12,10 +12,10 @@ pub enum PacketInfo {
 	Error(String),
 }
 
-pub fn handle_packet(p: Packet, quiet: bool) -> PacketInfo {
+pub fn handle_packet(p: Packet, quiet: bool, last_file: &mut Option<String>) -> PacketInfo {
 	let res = match p {
-		Packet::FileCreate { path, content } => handle_filecreate(path, content, quiet),
-		Packet::FileAppend { path, content } => handle_fileappend(path, content),
+		Packet::FileCreate { path, content } => handle_filecreate(path, content, quiet, last_file),
+		Packet::FileAppend { path, content } => handle_fileappend(path, content, last_file),
 		Packet::DirectoryCreate { path } => handle_dircreate(path, quiet),
 		Packet::Done => return PacketInfo::Stop,
 	};
@@ -26,20 +26,25 @@ pub fn handle_packet(p: Packet, quiet: bool) -> PacketInfo {
 	}
 }
 
-fn handle_filecreate(path: String, content: Vec<u8>, quiet: bool) -> Result<(), String> {
+fn handle_filecreate(path: String, content: Vec<u8>, quiet: bool, last_file: &mut Option<String>) -> Result<(), String> {
 	if !quiet { println!("creating file: {}", &path); }
 
-	let pbuf = PathBuf::from(path);
+	*last_file = Some(path.clone());
+
+	let pbuf = PathBuf::from(path.clone());
 	if !is_allowed(&pbuf)? {
 		return Err(format!("Creating File is not allowed: {:?}", pbuf));
 	}
+	if pbuf.exists() { return Err(format!("Trying to create File, which already exists: {}", &path)); }
 	let mut f = File::create(pbuf)
 		.map_err(|x| format!("Failed creating File: {}", x))?;
 	f.write_all(&content[..])
 		.map_err(|x| format!("Failed writing to File: {}", x))
 }
 
-fn handle_fileappend(path: String, content: Vec<u8>) -> Result<(), String> {
+fn handle_fileappend(path: String, content: Vec<u8>, last_file: &Option<String>) -> Result<(), String> {
+	if &Some(path.clone()) != last_file { return Err(format!("trying to append to File, which wasn't created last: {}", &path)); }
+
 	let pbuf = PathBuf::from(path);
 	if !is_allowed(&pbuf)? {
 		return Err(format!("Appending to File is not allowed: {:?}", pbuf));
@@ -53,10 +58,11 @@ fn handle_fileappend(path: String, content: Vec<u8>) -> Result<(), String> {
 fn handle_dircreate(path: String, quiet: bool) -> Result<(), String> {
 	if !quiet { println!("creating dir:  {}", &path); }
 
-	let pbuf = PathBuf::from(path);
+	let pbuf = PathBuf::from(path.clone());
 	if !is_allowed(&pbuf)? {
 		return Err(format!("Creating Directory is not allowed: {:?}", pbuf));
 	}
+	if pbuf.exists() { return Err(format!("Trying to create Dir, which already exists: {}", &path)); }
 
 	create_dir(pbuf)
 		.map_err(|x| format!("Failed creating Directory: {}", x))
